Move to Funcky.Internal
